Update some tests and correct code smells
