+++ src/main/java/com/upc/gessi/qrapids/app/domain/models/Project.java
@@ -1,10 +1,6 @@
 package com.upc.gessi.qrapids.app.domain.models;
 
-import org.hibernate.annotations.LazyCollection;
-import org.hibernate.annotations.LazyCollectionOption;
-
 import javax.persistence.*;
-import java.util.List;
 
 @Entity
 @Table(name = "project")
