+++ src/main/java/com/upc/gessi/qrapids/app/domain/repositories/Project/ProjectRepository.java
@@ -1,9 +1,7 @@
 package com.upc.gessi.qrapids.app.domain.repositories.Project;
 
 import com.upc.gessi.qrapids.app.domain.models.Project;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
 
 
 public interface ProjectRepository extends CrudRepository<Project, Long> {
