+++ src/test/java/com/upc/gessi/qrapids/app/presentation/rest/services/ProductsTest.java
@@ -70,59 +70,6 @@ public class ProductsTest {
     }
 
     @Test
-    public void getProjects() throws Exception {
-        Long projectId = 1L;
-        String projectExternalId = "test";
-        String projectName = "Test";
-        String projectDescription = "Test project";
-        boolean active = true;
-        String projectBacklogId = "999";
-        DTOProject dtoProject = new DTOProject(projectId, projectExternalId, projectName, projectDescription, null, active, projectBacklogId);
-        List<DTOProject> dtoProjectList = new ArrayList<>();
-        dtoProjectList.add(dtoProject);
-
-        when(productsController.getProjects()).thenReturn(dtoProjectList);
-
-        // Perform request
-        RequestBuilder requestBuilder = MockMvcRequestBuilders
-                .get("/api/projects");
-
-        this.mockMvc.perform(requestBuilder)
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$", hasSize(1)))
-                .andExpect(jsonPath("$[0].id", is(projectId.intValue())))
-                .andExpect(jsonPath("$[0].externalId", is(projectExternalId)))
-                .andExpect(jsonPath("$[0].name", is(projectName)))
-                .andExpect(jsonPath("$[0].description", is(projectDescription)))
-                .andExpect(jsonPath("$[0].logo", is(nullValue())))
-                .andExpect(jsonPath("$[0].active", is(active)))
-                .andExpect(jsonPath("$[0].backlogId", is(projectBacklogId)))
-                .andDo(document("projects/all",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        responseFields(
-                                fieldWithPath("[].id")
-                                        .description("Project identifier"),
-                                fieldWithPath("[].externalId")
-                                        .description("Project external identifier"),
-                                fieldWithPath("[].name")
-                                        .description("Project name"),
-                                fieldWithPath("[].description")
-                                        .description("Project description"),
-                                fieldWithPath("[].logo")
-                                        .description("Project logo file"),
-                                fieldWithPath("[].active")
-                                        .description("Is an active project?"),
-                                fieldWithPath("[].backlogId")
-                                        .description("Project identifier in the backlog"))
-                ));
-
-        // Verify mock interactions
-        verify(productsController, times(1)).getProjects();
-        verifyNoMoreInteractions(productsController);
-    }
-
-    @Test
     public void getProducts() throws Exception {
         Long projectId = 1L;
         String projectExternalId = "test";
@@ -272,170 +219,6 @@ public class ProductsTest {
 
         // Verify mock interactions
         verify(productsController, times(1)).getProductById(productId.toString());
-        verifyNoMoreInteractions(productsController);
-    }
-
-    @Test
-    public void getProjectById() throws Exception {
-        Long projectId = 1L;
-        String projectExternalId = "test";
-        String projectName = "Test";
-        String projectDescription = "Test project";
-        boolean active = true;
-        String projectBacklogId = "999";
-        DTOProject dtoProject = new DTOProject(projectId, projectExternalId, projectName, projectDescription, null, active, projectBacklogId);
-
-        when(productsController.getProjectById(projectId.toString())).thenReturn(dtoProject);
-
-        // Perform request
-        RequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                .get("/api/projects/{id}", projectId);
-
-        this.mockMvc.perform(requestBuilder)
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", is(projectId.intValue())))
-                .andExpect(jsonPath("$.externalId", is(projectExternalId)))
-                .andExpect(jsonPath("$.name", is(projectName)))
-                .andExpect(jsonPath("$.description", is(projectDescription)))
-                .andExpect(jsonPath("$.logo", is(nullValue())))
-                .andExpect(jsonPath("$.active", is(active)))
-                .andExpect(jsonPath("$.backlogId", is(projectBacklogId)))
-                .andDo(document("projects/single",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        pathParameters(
-                                parameterWithName("id")
-                                        .description("Project identifier")
-                        ),
-                        responseFields(
-                                fieldWithPath("id")
-                                        .description("Project identifier"),
-                                fieldWithPath("externalId")
-                                        .description("Project external identifier"),
-                                fieldWithPath("name")
-                                        .description("Project name"),
-                                fieldWithPath("description")
-                                        .description("Project description"),
-                                fieldWithPath("logo")
-                                        .description("Project logo file"),
-                                fieldWithPath("active")
-                                        .description("Is an active project?"),
-                                fieldWithPath("backlogId")
-                                        .description("Project identifier in the backlog"))
-                ));
-
-        // Verify mock interactions
-        verify(productsController, times(1)).getProjectById(projectId.toString());
-        verifyNoMoreInteractions(productsController);
-    }
-
-    @Test
-    public void updateProject() throws Exception {
-        Long projectId = 1L;
-        String projectExternalId = "test";
-        String projectName = "Test";
-        String projectDescription = "Test project";
-        String projectBacklogId = "999";
-// getResource() : The name of a resource is a '/'-separated path name that identifies the resource.
-        URL projectImageUrl = QrapidsApplication.class.getClassLoader().getResource("static" + "/" + "icons" + "/" + "projectDefault.jpg");
-        File file = new File(projectImageUrl.getPath());
-        MockMultipartFile logoMultipartFile = new MockMultipartFile("logo", "logo.jpg", "image/jpeg", Files.readAllBytes(file.toPath()));
-
-        DTOProject dtoProject = new DTOProject(projectId, projectExternalId, projectName, projectDescription, logoMultipartFile.getBytes(), true, projectBacklogId);
-
-        when(productsController.checkProjectByName(projectId, projectName)).thenReturn(true);
-
-        // Perform request
-        RequestBuilder requestBuilder = MockMvcRequestBuilders
-                .multipart("/api/projects/{id}", projectId)
-                .file(logoMultipartFile)
-                .param("externalId", projectExternalId)
-                .param("name", projectName)
-                .param("description", projectDescription)
-                .param("backlogId", projectBacklogId)
-                .with(new RequestPostProcessor() {
-                    @Override
-                    public MockHttpServletRequest postProcessRequest(MockHttpServletRequest request) {
-                        request.setMethod("PUT");
-                        return request;
-                    }
-                });
-
-        this.mockMvc.perform(requestBuilder)
-                .andExpect(status().isOk())
-                .andDo(document("projects/update",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        requestParameters(
-                                parameterWithName("externalId")
-                                        .description("Project external identifier"),
-                                parameterWithName("name")
-                                        .description("Project name"),
-                                parameterWithName("description")
-                                        .description("Project description"),
-                                parameterWithName("backlogId")
-                                        .description("Project identifier in the backlog")),
-                        requestParts(
-                                partWithName("logo")
-                                        .description("Project logo file")
-                        )
-                ));
-
-        // Verify mock interactions
-        verify(productsController, times(1)).checkProjectByName(projectId, projectName);
-
-        ArgumentCaptor<DTOProject> argument = ArgumentCaptor.forClass(DTOProject.class);
-        verify(productsController, times(1)).updateProject(argument.capture());
-        assertEquals(dtoProject.getId(), argument.getValue().getId());
-        assertEquals(dtoProject.getExternalId(), argument.getValue().getExternalId());
-        assertEquals(dtoProject.getName(), argument.getValue().getName());
-        assertEquals(dtoProject.getDescription(), argument.getValue().getDescription());
-        assertEquals(dtoProject.getActive(), argument.getValue().getActive());
-        assertEquals(dtoProject.getExternalId(), argument.getValue().getExternalId());
-
-        verifyNoMoreInteractions(productsController);
-    }
-
-    @Test
-    public void updateProjectNameAlreadyExists() throws Exception {
-        Long projectId = 1L;
-        String projectExternalId = "test";
-        String projectName = "Test";
-        String projectDescription = "Test project";
-        String projectBacklogId = "999";
-// getResource() : The name of a resource is a '/'-separated path name that identifies the resource.
-        URL projectImageUrl = QrapidsApplication.class.getClassLoader().getResource("static" + "/" + "icons" + "/" + "projectDefault.jpg");
-        File file = new File(projectImageUrl.getPath());
-        MockMultipartFile logoMultipartFile = new MockMultipartFile("logo", "logo.jpg", "image/jpeg", Files.readAllBytes(file.toPath()));
-
-        when(productsController.checkProjectByName(projectId, projectName)).thenReturn(false);
-
-        // Perform request
-        RequestBuilder requestBuilder = MockMvcRequestBuilders
-                .multipart("/api/projects/{id}", projectId)
-                .file(logoMultipartFile)
-                .param("externalId", projectExternalId)
-                .param("name", projectName)
-                .param("description", projectDescription)
-                .param("backlogId", projectBacklogId)
-                .with(new RequestPostProcessor() {
-                    @Override
-                    public MockHttpServletRequest postProcessRequest(MockHttpServletRequest request) {
-                        request.setMethod("PUT");
-                        return request;
-                    }
-                });
-
-        this.mockMvc.perform(requestBuilder)
-                .andExpect(status().isConflict())
-                .andDo(document("projects/update-error",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint())
-                ));
-
-        // Verify mock interactions
-        verify(productsController, times(1)).checkProjectByName(projectId, projectName);
-
         verifyNoMoreInteractions(productsController);
     }
